Code reordered, documentation pairs with code, same ordering.
This is made nice for us at release time when DZIL moves it all the way
to the end.

The primary reasons to keep POD after __END__ are for readability,
as it suffers when you don't have a syntax highlighter or code folding
available.

However, when developing stuff, we have these facilities, and having the
documentation closer to the code is more advantageous as it encourages
keeping it up to date.

PodPurler makes this a legitamate possibility